navigation component updated
