+++ src/components/layout/navigation/navigation.module.css
@@ -1,0 +1,288 @@
+.root {
+  position: fixed;
+  z-index: 50;
+  display: flex;
+  height: 5rem;
+  width: 100vw;
+  align-items: center;
+  justify-content: space-between;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  transition-property: all;
+  transition-duration: 300ms;
+
+  @media (min-width: 1024px) {
+    padding-left: 2.5rem;
+    padding-right: 2.5rem;
+  }
+
+  @media (min-width: 1280px) {
+    padding-left: 4rem;
+    padding-right: 4rem;
+  }
+
+  &.scrolled {
+    background-color: var(--background);
+
+    @media (prefers-color-scheme: dark) {
+      color: #ffffff;
+    }
+  }
+
+  &.top {
+    background-color: transparent;
+
+    @media (prefers-color-scheme: dark) {
+      color: #ffffff;
+    }
+  }
+
+  &.hidden {
+    transform: translateY(-100%);
+    opacity: 0;
+  }
+
+  &.visible {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+.container {
+  margin-left: auto;
+  margin-right: auto;
+  display: flex;
+  width: 100%;
+  max-width: 1920px;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+
+  @media (min-width: 1024px) {
+    gap: 1.5rem;
+  }
+}
+
+.email {
+  font-size: 0.75rem;
+
+  @media (min-width: 768px) {
+    font-size: 0.875rem;
+  }
+
+  @media (min-width: 1024px) {
+    font-size: 1rem;
+  }
+}
+
+.return-link {
+  font-family: var(--font-heading);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+
+  @media (min-width: 1024px) {
+    font-size: 1.125rem;
+  }
+}
+
+.return-icon {
+  cursor: pointer;
+  transition-property: transform;
+  transition-duration: 200ms;
+}
+
+.return-link:hover .return-icon {
+  transform: translateX(-0.25rem);
+}
+
+.controls {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.cvWrapper {
+  display: none;
+
+  @media (min-width: 460px) {
+    display: flex;
+  }
+}
+
+.cv {
+  display: none;
+  @media (min-width: 460px) {
+    display: flex;
+    font-family: var(--font-heading);
+    color: var(--foreground);
+    border-radius: 9999px;
+    background-color: #ffffff;
+    padding-left: 0.75rem;
+    padding-right: 0.75rem;
+    font-size: 0.875rem;
+    transition-property: all;
+    transition-duration: 300ms;
+  }
+
+  @media (min-width: 1024px) {
+    padding-left: 1rem;
+    padding-right: 1rem;
+    font-size: 1rem;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    color: var(--background);
+  }
+}
+
+.hamburger {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  gap: 0.25rem;
+  cursor: pointer;
+
+  @media (min-width: 1024px) {
+    display: none;
+  }
+}
+
+.line {
+  background-color: var(--foreground);
+  z-index: 50;
+  height: 3px;
+  border-radius: 9999px;
+  transition-property: all;
+  transition-duration: 300ms;
+}
+
+.line1 {
+  composes: line;
+  width: 1.5rem;
+  transform: translateY(0) rotate(0);
+
+  &.open {
+    width: 1.5rem;
+    transform: translateY(7px) rotate(45deg);
+  }
+}
+
+.line2 {
+  composes: line;
+  width: 2rem;
+  opacity: 1;
+
+  &.open {
+    opacity: 0;
+  }
+}
+
+.line3 {
+  composes: line;
+  width: 1.25rem;
+  transform: translateY(0) rotate(0);
+
+  &.open {
+    width: 1.5rem;
+    transform: translateY(-7px) rotate(-45deg);
+  }
+}
+
+.menu {
+  transform: translateX(100%);
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: 20;
+  height: 100vh;
+  width: 70%;
+  background-color: var(--whitebg);
+  transition-property: transform;
+  transition-duration: 300ms;
+  transition-timing-function: ease-in-out;
+
+  @media (prefers-color-scheme: dark) {
+    background-color: var(--background);
+  }
+
+  &.open {
+    transform: translateX(0);
+  }
+
+  @media (min-width: 768px) {
+    width: 45%;
+  }
+
+  @media (min-width: 1024px) {
+    transform: translateX(0);
+    position: static;
+    height: auto;
+    width: auto;
+    display: flex;
+    align-items: center;
+    gap: 2.5rem;
+    font-size: 1.25rem;
+    transform: none;
+    background-color: transparent;
+  }
+}
+
+.list {
+  font-family: var(--font-heading);
+  position: absolute;
+  top: 6rem;
+  left: 2.5rem;
+  z-index: 100;
+  display: flex;
+  width: 75%;
+  flex-direction: column;
+  gap: 2rem;
+  font-size: 1.5rem;
+  font-weight: 500;
+
+  @media (min-width: 640px) {
+    left: 5rem;
+    gap: 3rem;
+    font-size: 1.875rem;
+  }
+
+  @media (min-width: 1024px) {
+    position: static;
+    width: auto;
+    flex-direction: row;
+    gap: 2.5rem;
+    font-size: 1.25rem;
+    font-weight: 400;
+  }
+}
+
+.social {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+
+  &:hover .arrow {
+    @media (min-width: 1024px) {
+      transform: translate(2px, -2px) scale(1.1);
+    }
+  }
+}
+
+.arrow {
+  height: 1.75rem;
+  width: 1.75rem;
+
+  @media (min-width: 1024px) {
+    margin-top: 0.25rem;
+    height: 1rem;
+    width: 1rem;
+    transition-property: transform;
+    transition-duration: 200ms;
+  }
+}
