+++ src/components/sections/hero/hero.module.css
@@ -84,11 +84,6 @@
   height: 8rem;
   width: 8rem;
 
-  @media (prefers-color-scheme: dark) {
-    border: none;
-    background-image: linear-gradient(to top right, #1a1a1a, #2d2d2d);
-  }
-
   @media (min-width: 640px) {
     height: 9rem;
     width: 9rem;
@@ -101,6 +96,11 @@
   @media (min-width: 1280px) {
     height: 11rem;
     width: 11rem;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    border: none;
+    background-image: linear-gradient(to top right, #1a1a1a, #2d2d2d);
   }
 }
