+++ src/components/sections/skills/skills.module.css
@@ -6,10 +6,6 @@
   padding-top: 2.5rem;
   padding-bottom: 5rem;
 
-  @media (prefers-color-scheme: dark) {
-    background-color: #1a1a1a;
-  }
-
   @media (min-width: 768px) {
     padding-top: 3.5rem;
     padding-bottom: 6rem;
@@ -29,6 +25,10 @@
   @media (min-width: 1536px) {
     border-bottom-left-radius: 70px;
     border-bottom-right-radius: 70px;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #1a1a1a;
   }
 }
 
@@ -135,14 +135,6 @@
   padding: 1rem;
   transition: all 300ms;
 
-  @media (prefers-color-scheme: dark) {
-    background-color: #2c2c2c;
-  }
-
-  @media (min-width: 640px) {
-    padding: 1.5rem;
-  }
-
   &:hover {
     transform: scale(0.95);
   }
@@ -161,6 +153,14 @@
   &:hover .badge {
     transform: translateY(0);
     opacity: 1;
+  }
+
+  @media (min-width: 640px) {
+    padding: 1.5rem;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #2c2c2c;
   }
 }
